Coworking Session from yesterday.
