+++ src/main/java/AccountLogin/LoginPresenter.java
@@ -1,8 +1,19 @@
 package AccountLogin;
 
+import AccountCreation.Account;
+import Swipe.Screen.EmptyScreen;
+import Swipe.Screen.SwipeScreen;
+import Swipe.Screen.SwiperController;
+import Swipe.Screen.SwiperPresenterFormatter;
+import Swipe.SwiperInputBoundary;
+import Swipe.SwiperInteractor;
+import Swipe.SwiperPresenter;
 import UI.LoginUI;
+import UI.NavigationUI;
 import UI.SwiperUI;
 import UI.ViewModel;
+import data.persistency.UserDatabase;
+import matching.MatchingAlgorithm;
 
 import javax.swing.*;
 
@@ -20,10 +31,15 @@ public class LoginPresenter implements LoginOutBoundary{
         if (responseModel.getLoginStatus()) {
             loginUI.loginSuccessMechanism();
             // TODO: go to swiperUI of this user's account.
-            JFrame frame = new JFrame("FrameDemo");
-            frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-            frame.pack();
-            frame.setVisible(true);
+
+
+            //Open SwiperUI frame
+            swipePage();
+
+
+            //Open navigation frame
+            NavigationUI navigationUI = new NavigationUI();
+            navigationUI.build();
 
             loginUI.closeLoginUI();
         } else {
