+++ src/main/java/data/persistency/UserDataAccessInterface.java
@@ -5,6 +5,5 @@ import AccountCreation.Account;
 import java.util.ArrayList;
 
 public interface UserDataAccessInterface {
-    void save();
-    ArrayList<Account> load();
+
 }
